`readder` rea

+++ src/backend/audio/Music.java
@@ -1,10 +1,16 @@
 package backend.audio;
 
 import it.sauronsoftware.jave.*;
+import javazoom.jl.converter.*;
+import javazoom.jl.decoder.JavaLayerException;
 
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.IOException;
 import java.nio.file.Files;
+
+import app.CLI;
+import backend.audioutil.PlayerProgress;
 
 public class Music {
   private Music() {
@@ -25,35 +31,41 @@ public class Music {
     return sb.toString();
   }
 
-  /**
-   * Plans:
-   * To copy the MP3 over to the temp folder, and then convert it to WAV.
-   * Then keep a record of that conversion so we know.
-   * @param f
-   * @return
-   * \    pm.p.setProperty(f2.getAbsolutePath(), f.getName());
-    pm.save();
-   */
-  public static File convert(File f) {
+  private static File secondaryConvert(File mp3, File dest) {
+    //assert mp3.getName().endsWith(".mp3");
+    Converter c = new Converter();
+    PlayerProgress s = new PlayerProgress();
+    s.start();
+    try {
+      System.out.print(mp3.getAbsolutePath() + "\n" + dest.getAbsolutePath());
+      c.convert(mp3.getAbsolutePath(), dest.getAbsolutePath(), s);
+    } catch (Exception e) {
+      e.printStackTrace();
+    }
+    s.f.dispose();
+    return dest;
+  }
+
+  public static File externConvert(File f) {
     String fName = f.getName();
-    // remove all unicode characters
     fName = fName.replaceAll("[^a-zA-Z0-9\\s]", "");
-    fName = fName.replaceAll("\\s+", "");
+    fName = fName.replace(" ", "");
+    fName += ".mp3";   
     File wav = new File(GlobalVars.ITEM_DIR + "\\" + fName.replace(".mp3", "") + ".wav");
-    File f2 = new File(GlobalVars.ITEM_DIR + "\\" + randomFileName() + ".mp3");
+    File f2 = new File(GlobalVars.ITEM_DIR + "\\" + fName);
     if (wav.exists()) {
-      System.out.println("WAV file already exists, skipping conversion");
+      System.out.println("WAV file already exists, skipping conversion" + wav.getAbsolutePath());
       return wav;
     } else {
       try {
-        Files.copy(f.toPath(), f2.toPath());
+        if (!f2.exists())
+          Files.copy(f.toPath(), f2.toPath());
       } catch (IOException e) {
         e.printStackTrace();
-      } finally {
-        // DO NOTHING
       }
     }
-    System.out.println("Converting " + f.getAbsolutePath() + " to " + wav.getAbsolutePath());
+    System.out.println(f2.getAbsolutePath());
+    //System.out.println("Converting " + f.getAbsolutePath() + " to " + wav.getAbsolutePath());
     AudioAttributes audio = new AudioAttributes();
     audio.setCodec("pcm_s16le");
     audio.setBitRate(320000);
@@ -65,13 +77,67 @@ public class Music {
     Encoder encoder = new Encoder();
     try {
       encoder.encode(f2, wav, attrs);
-    } catch (IllegalArgumentException e) {
+    } catch (it.sauronsoftware.jave.InputFormatException e) {
       e.printStackTrace();
-    } catch (InputFormatException e) {
-      e.printStackTrace();
-    } catch (EncoderException e) {
+      return secondaryConvert(f2, wav);
+    } catch (IllegalArgumentException | it.sauronsoftware.jave.EncoderException e) {
       e.printStackTrace();
     }
+
     return wav;
+  }
+
+  /**
+   * Plans:
+   * To copy the MP3 over to the temp folder, and then convert it to WAV.
+   * Then keep a record of that conversion so we know.
+   * 
+   * @return
+   *         \ pm.p.setProperty(f2.getAbsolutePath(), f.getName());
+   *         pm.save();
+   */
+  public static File convert(File f) {
+    String fName = f.getName();
+    // remove all unicode characters
+    fName = fName.replaceAll("[^a-zA-Z0-9\\s]", "");  
+    fName = fName.replace(" ", "");
+    fName += ".mp3"; 
+    File wav = new File(GlobalVars.ITEM_DIR + "\\" + fName.replace(".mp3", "") + ".wav");
+    File f2 = new File(GlobalVars.ITEM_DIR + "\\" + fName);
+    if (wav.exists()) {
+      System.out.println("WAV file already exists, skipping conversion" + wav.getAbsolutePath());
+      return wav;
+    } else {
+      try {
+        if (!f2.exists())
+          Files.copy(f.toPath(), f2.toPath());
+      } catch (IOException e) {
+        e.printStackTrace();
+      }
+    }
+    System.out.println("Converting " + f.getAbsolutePath() + " to " + wav.getAbsolutePath());
+    return secondaryConvert(f2, wav);
+    /*
+    AudioAttributes audio = new AudioAttributes();
+    // dynanmically pick the codec to use to decode mp3 to wav
+    audio.setCodec("pcm_s16le");
+    audio.setBitRate(320000);
+    audio.setChannels(2);
+    audio.setSamplingRate(44100);
+    EncodingAttributes attrs = new EncodingAttributes();
+    attrs.setFormat("wav");
+    attrs.setAudioAttributes(audio);
+    Encoder encoder = new Encoder();
+    try {
+      encoder.encode(f2, wav, attrs);
+    } catch (it.sauronsoftware.jave.InputFormatException e) {
+      e.printStackTrace();
+      return secondaryConvert(f2, wav);
+    } catch (IllegalArgumentException | it.sauronsoftware.jave.EncoderException e) {
+      e.printStackTrace();
+    }
+
+    return wav;
+    */
   }
 }
